requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
openpyxl==3.1.2
flet